+++ src/components/Icon/index.js
@@ -3,6 +3,7 @@ export * from './Icon';
 export * from './icons/IconAttach';
 export * from './icons/IconAttention';
 export * from './icons/IconArrow';
+export * from './icons/IconBell';
 export * from './icons/IconBox';
 export * from './icons/IconBoxNoData';
 export * from './icons/IconBucket';