+++ src/index.js
@@ -1,4 +1,6 @@
 export * from './components/Alert';
+export * from './components/AppHeader';
+export * from './components/AppLayout';
 export * from './components/BaseModal';
 export * from './components/Breadcrumbs';
 export * from './components/Button';